Debian packaging for kopano-webapp-plugin-desktopnotifications
This creates a Debian package for the Kopano Webapp desktopnotifications
plugin. It does not use the Ant build system that upstream does, but
a few simple shell commands to build things. This reduces the complexity
of the package a lot, but the drawback is that the rules file must be
validated every time a new upstream version is imported.

Many thanks to Carsten Schoenert for his work on packaging the Files
plugin, from which much of this was borrowed.